+++ index.js
@@ -1,6 +1,8 @@
-const pos = require('./dist/pos');
+const PosIntegrado = require('./dist/PosIntegrado');
+const PosAutoservicio = require('./dist/PosAutoservicio')
 
 const Transbank = {};
-Transbank.POS = pos;
+Transbank.POSIntegrado = PosIntegrado;
+Transbank.POSAutoservicio = PosAutoservicio;
 
 module.exports = Transbank;
